latitude longitude validation check changes2

+++ fir/forms.py
@@ -41,10 +41,10 @@ class FirForm(forms.ModelForm):
         if cd.get('dri_lic_date_issu') > cd.get('dri_lic_date_upto'):
             self.add_error('dri_lic_date_upto', "Driver License Validity cannot be before Issued Date")
             
-        if int(cd.get('LATITUDE')) > 29 or int(cd.get('LATITUDE')) < 28:
+        if float(cd.get('LATITUDE')) > 29 or float(cd.get('LATITUDE')) < 28:
              self.add_error('LATITUDE', "Check Value of Latitude")
              
-        if int(cd.get('LONGITUDE')) > 78 or int(cd.get('LONGITUDE')) < 76:
+        if float(cd.get('LONGITUDE')) > 78 or float(cd.get('LONGITUDE')) < 76:
              self.add_error('LONGITUDE', "Check Value of Longitude") 
              
         tim1 = cd.get('TIME_OCC')[:2]
